ch2: refactoring
